WIP: Use the SWIG doxygen option to replace JavaDoc.i
Enables SWIG to automatically produce comments in the Java code based
on processed C++ Doxygen comments. This remove the to manually keep
the JavaDoc.i file up to date and processing, and code maintenance
involved.

TODO: The SWIG process is producing a fair number of warnings related
to invalid "@see" commands, reference and characters.